Syncing with internal repo for Open62541 backend
