+++ Common/CMakeLists.txt
@@ -2,4 +2,5 @@
 
 add_library (Common OBJECT
 	src/ASUtils.cpp
+
 	)
