+++ Configuration/designToConfigurator.xslt
@@ -161,12 +161,14 @@ xsi:schemaLocation="http://www.w3.org/1999/XSL/Transform schema-for-xslt20.xsd "
 				nm->getTypeNodeId(AddressSpace::ASInformationModel::<xsl:value-of select="fnc:typeNumericId($class)"/>), 
 				nm, 
 				<xsl:value-of select="@class"/>config);
+		#ifndef BACKEND_OPEN62541
 		UaStatus s = nm->addNodeAndReference( parentNodeId, asItem, OpcUaId_HasComponent);
 		if (!s.isGood())
 		{
 			std::cout &lt;&lt; "While addNodeAndReference from " &lt;&lt; parentNodeId.toString().toUtf8() &lt;&lt; " to " &lt;&lt; asItem-&gt;nodeId().toString().toUtf8() &lt;&lt; " : " &lt;&lt; std::endl;
 			ASSERT_GOOD(s);
 		}
+		#endif
 		<xsl:if test="fnc:classHasDeviceLogic(/,$class)='true'">
 		Device::<xsl:value-of select="fnc:DClassName(@name)"/> *dItem = new Device::<xsl:value-of select="fnc:DClassName(@name)"/> (config
 		<xsl:if test="fnc:getCountParentClassesAndRoot(/,@name)=1">
@@ -208,8 +210,11 @@ xsi:schemaLocation="http://www.w3.org/1999/XSL/Transform schema-for-xslt20.xsd "
 	#include &lt;Configurator.h&gt;
 	#include &lt;Configuration.hxx&gt;
 
+	#ifndef BACKEND_OPEN62541
 	#include &lt;meta.h&gt;
+	#endif
 
+	#include &lt;LogIt.h&gt;
 	
 <!-- *************************************************** -->
 <!-- HEADERS OF ALL DECLARED CLASSES ******************* -->
@@ -250,20 +255,22 @@ xsi:schemaLocation="http://www.w3.org/1999/XSL/Transform schema-for-xslt20.xsd "
 	} 
 	catch (xsd::cxx::tree::parsing&lt;char&gt; &amp;exception)
 	{
-        std::cout &lt;&lt; "Failed when trying to open the configuration, with general error message: " &lt;&lt; exception.what() &lt;&lt; std::endl;
+        LOG(Log::ERR) &lt;&lt; "Configuration: Failed when trying to open the configuration, with general error message: " &lt;&lt; exception.what();
 		BOOST_FOREACH( const xsd::cxx::tree::error&lt;char&gt; &amp;error, exception.diagnostics() )
 		{
-			std::cout &lt;&lt; "Problem at " &lt;&lt; error.id() &lt;&lt; ":" &lt;&lt; error.line() &lt;&lt; ": " &lt;&lt; error.message() &lt;&lt; std::endl;
+			LOG(Log::ERR) &lt;&lt; "Configuration: Problem at " &lt;&lt; error.id() &lt;&lt; ":" &lt;&lt; error.line() &lt;&lt; ": " &lt;&lt; error.message();
 		}
-	    return false;
+	    throw std::runtime_error("Configuration: failed to load configuration. The exact problem description should have been logged.");
 	}
 	
 	
 	UaNodeId rootNode = UaNodeId(OpcUaId_ObjectsFolder, 0);
 	Device::DRoot *deviceRoot = Device::DRoot::getInstance();
-		
+
+	#ifndef BACKEND_OPEN62541
 	configureMeta( *theConfiguration.get(), nm, rootNode );	
-		
+	#endif
+	
 	<xsl:for-each select="/d:design/d:root/d:hasobjects[@instantiateUsing='configuration']">
 	BOOST_FOREACH(const Configuration::<xsl:value-of select="@class"/> &amp; <xsl:value-of select="@class"/>config, theConfiguration-&gt;<xsl:value-of select="@class"/>())
 	{
@@ -327,14 +334,14 @@ xsi:schemaLocation="http://www.w3.org/1999/XSL/Transform schema-for-xslt20.xsd "
 <!-- *************************************************** -->
 	void unlinkAllDevices (AddressSpace::ASNodeManager *nm)
 	{
+	        unsigned int totalObjectsNumber = 0;
 		<xsl:for-each select="/d:design/d:class">
 		<xsl:if test="fnc:classHasDeviceLogic(/,@name)='true'">
 		{
 			std::vector&lt; AddressSpace::<xsl:value-of select="fnc:ASClassName(@name)"/> * &gt; objects;
 			std::string pattern (".*");
 			nm->findAllByPattern&lt;AddressSpace::<xsl:value-of select="fnc:ASClassName(@name)"/>&gt; ( nm-&gt;getNode(UaNodeId(OpcUaId_ObjectsFolder, 0)), OpcUa_NodeClass_Object, pattern, objects);
-			PRINT("objects.size=");
-			PRINT(objects.size());
+			totalObjectsNumber += objects.size();
 			BOOST_FOREACH(AddressSpace::<xsl:value-of select="fnc:ASClassName(@name)"/> *a, objects)
 			{
 				a-&gt;unlinkDevice();
@@ -342,6 +349,7 @@ xsi:schemaLocation="http://www.w3.org/1999/XSL/Transform schema-for-xslt20.xsd "
 		}
 		</xsl:if>
 		</xsl:for-each>
+		LOG(Log::INF) &lt;&lt; "Total number of unlinked objects: " &lt;&lt; totalObjectsNumber;
 	}
 	
 	</xsl:template>
