+++ Documentation/DesignManual.html
@@ -1,1 +1,1 @@
-(C) CERN 2015. All rights not expressly granted are reserved. <br><br><br>The Generic OPC UA Server Framework<br><br>Design File Manual<br><br><h1>Class</h1><br><br><br><br><h1>Cachevariable</h1><h2>Attributes of cachevariable</h2><h3>addressSpaceWrite</h3><h3>makeSetGet</h3>This attribute is deprecated and will be completely removed soon.<br><br><h3>dataType</h3><br><br><h3>initializeWith</h3>Determines what the cachevariable will be initialized with before any custom code or any OPC UA client write is effective.<br><ul><li>When configuration, an appropriate entry is made in the Configuration.xsd file which in turn adds s<br></li></ul><br>initialValue<br><br><h3>initialStatus</h3>When initializeWith=valueAndStatus, this attribute gives the initial status of the cachevariable.<br>When initializeWith is different, this attribute is ignored.<br><br><h3>isKey</h3><h3>nullPolicy</h3><ul><li>When nullForbidden, the cachevariable is never allowed to contain NULL. This applies to setters (Device Logic can't set NULL), OPC UA Client write operations (write operation carrying NULL will be denied) and initialization (when initializeFrom=valueAndStatus, initialValue is mandatory to be given). Also, thanks to this setting a short getter will be created.</li><li>When nullAllowed, there is no restriction towards NULL.<br></li></ul><br><h2>Relations between nullPolicy, initializeWith and initialValue attributes</h2><br><table style="text-align: left; width: 100%;" border="1" cellpadding="2" cellspacing="2"><tbody><tr><td style="vertical-align: top;"><span style="font-weight: bold;">nullPolicy</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">initializeWith</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">initialValue</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">Comment</span><br></td></tr><tr><td style="vertical-align: top;">nullAllowed<br></td><td style="vertical-align: top;">configuration<br></td><td style="vertical-align: top;">NOT RELEVANT<br></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullAllowed<br></td><td style="vertical-align: top;">valueAndStatus<br></td><td style="vertical-align: top;"><ul><li>If not present, the value is initialized as NULL.</li><li>If present, the value is initialized with the attribute contents.<br></li></ul></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullForbidden<br></td><td style="vertical-align: top;">configuration<br></td><td style="vertical-align: top;">NOT RELEVANT<br></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullForbidden</td><td style="vertical-align: top;">valueAndStatus<br></td><td style="vertical-align: top;">MUST BE PRESENT<br></td><td style="vertical-align: top;"><br></td></tr></tbody></table><br>
+The Generic OPC UA Server Framework<br><br>Design File Manual<br><br><h1>Class</h1><br><br><br><br><h1>Cachevariable</h1><h2>Attributes of cachevariable</h2><h3>addressSpaceWrite</h3><h3>makeSetGet</h3>This attribute is deprecated and will be completely removed soon.<br><br><h3>dataType</h3><br><br><h3>initializeWith</h3>Determines what the cachevariable will be initialized with before any custom code or any OPC UA client write is effective.<br><ul><li>When configuration, an appropriate entry is made in the Configuration.xsd file which in turn adds s<br></li></ul><br>initialValue<br><br><h3>initialStatus</h3>When initializeWith=valueAndStatus, this attribute gives the initial status of the cachevariable.<br>When initializeWith is different, this attribute is ignored.<br><br><h3>isKey</h3><h3>nullPolicy</h3><ul><li>When nullForbidden, the cachevariable is never allowed to contain NULL. This applies to setters (Device Logic can't set NULL), OPC UA Client write operations (write operation carrying NULL will be denied) and initialization (when initializeFrom=valueAndStatus, initialValue is mandatory to be given). Also, thanks to this setting a short getter will be created.</li><li>When nullAllowed, there is no restriction towards NULL.<br></li></ul><br><h2>Relations between nullPolicy, initializeWith and initialValue attributes</h2><br><table style="text-align: left; width: 100%;" border="1" cellpadding="2" cellspacing="2"><tbody><tr><td style="vertical-align: top;"><span style="font-weight: bold;">nullPolicy</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">initializeWith</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">initialValue</span><br></td><td style="vertical-align: top;"><span style="font-weight: bold;">Comment</span><br></td></tr><tr><td style="vertical-align: top;">nullAllowed<br></td><td style="vertical-align: top;">configuration<br></td><td style="vertical-align: top;">NOT RELEVANT<br></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullAllowed<br></td><td style="vertical-align: top;">valueAndStatus<br></td><td style="vertical-align: top;"><ul><li>If not present, the value is initialized as NULL.</li><li>If present, the value is initialized with the attribute contents.<br></li></ul></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullForbidden<br></td><td style="vertical-align: top;">configuration<br></td><td style="vertical-align: top;">NOT RELEVANT<br></td><td style="vertical-align: top;"><br></td></tr><tr><td style="vertical-align: top;">nullForbidden</td><td style="vertical-align: top;">valueAndStatus<br></td><td style="vertical-align: top;">MUST BE PRESENT<br></td><td style="vertical-align: top;"><br></td></tr></tbody></table><br>
