+++ Documentation/GenericServer.html
@@ -15,7 +15,6 @@
 
 
 
-
   
   <meta charset="UTF-8"><title>QUASAR</title></head><body>
 
@@ -684,7 +683,9 @@ currentTemperature converging to the value.<br>
 <hr style="width: 100%; height: 2px;">
 Written by Piotr Nikiel &lt;piotr.nikiel@cern.ch&gt;<br>
 
-Report inconsistencies and bugs to the email above.<br>(C) CERN 2015. All rights not expressly granted are reserved.<br>
+Report inconsistencies and bugs to the email above.<br>
+
+The Generic OPC UA Server framework, (C) CERN 2014. <br>
 
 <br>
