+++ Extra/designToHonkyTonk.xslt
@@ -1,23 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- © Copyright CERN, 2015.                                                       -->
-<!-- All rights not expressly granted are reserved.                                -->
-<!-- This file is part of Quasar.                                                  -->
-<!--                                                                               -->
-<!-- Quasar is free software: you can redistribute it and/or modify                -->     
-<!-- it under the terms of the GNU Lesser General Public Licence as published by   -->     
-<!-- the Free Software Foundation, either version 3 of the Licence.                -->     
-<!-- Quasar is distributed in the hope that it will be useful,                     -->     
-<!-- but WITHOUT ANY WARRANTY; without even the implied warranty of                -->     
-<!--                                                                               -->
-<!-- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the                 -->
-<!-- GNU Lesser General Public Licence for more details.                           -->
-<!--                                                                               -->
-<!-- You should have received a copy of the GNU Lesser General Public License      -->
-<!-- along with Quasar.  If not, see <http://www.gnu.org/licenses/>                -->
-<!--                                                                               -->
-<!-- Created:   Oct 2014                                                           -->
-<!-- Authors:                                                                      -->
-<!--   Piotr Nikiel <piotr@nikiel.info>                                            -->
 <xsl:transform version="2.0" xmlns:xml="http://www.w3.org/XML/1998/namespace" 
 xmlns:xs="http://www.w3.org/2001/XMLSchema" 
 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
