+++ FrameworkInternals/original_files.txt
@@ -151,8 +151,8 @@ Directory Server install=create
 File CMakeLists.txt                         must_exist,must_be_versioned,install=ask_to_merge
 
 Directory Server/include install=create
-File BaseQuasarServer.h                    must_exist,must_be_versioned,md5=check,install=overwrite
-File QuasarServer.h                         must_exist,must_be_versioned,md5=check,install=copy_if_not_existing
+File BaseQuasarServer.h                     must_exist,must_be_versioned,md5=check,install=ask_to_merge
+File QuasarServer.h                         must_exist,must_be_versioned,md5=check,install=ask_to_merge
 File shutdown.h                             must_exist,must_be_versioned,md5=check,install=overwrite
 File serverconfigxml.h                      must_exist,must_be_versioned,md5=check,install=overwrite
 File opcserver.h                            must_exist,must_be_versioned,md5=check,install=overwrite
@@ -162,8 +162,8 @@ File QUASARFrameworkVersion.h               must_exist,must_be_versioned,md5=che
 
 Directory Server/src install=create
 File main.cpp                               must_exist,must_be_versioned,install=ask_to_merge
-File BaseQuasarServer.cpp                  must_exist,must_be_versioned,md5=check,install=overwrite
-File QuasarServer.cpp                       must_exist,must_be_versioned,md5=check,install=copy_if_not_existing
+File BaseQuasarServer.cpp                   must_exist,must_be_versioned,md5=check,install=ask_to_merge
+File QuasarServer.cpp                       must_exist,must_be_versioned,md5=check,install=ask_to_merge
 File serverconfigxml.cpp                    must_exist,must_be_versioned,md5=check,install=overwrite
 File shutdown.cpp                           must_exist,must_be_versioned,md5=check,install=overwrite
 File opcserver.cpp                          must_exist,must_be_versioned,md5=check,install=overwrite
@@ -187,8 +187,6 @@ File frameworkGUI.py                        deprecated
 File generated_files_list.cmake             deprecated,install=dont_touch
 File generated_headers_list.cmake           deprecated,install=dont_touch
 File transformDesign.sh               		deprecated,install=dont_touch
-File LICENSE.TXT                            must_exist,must_be_versioned,install=overwrite,md5=check,
-
 
 Directory Meta install=create,
 File CMakeLists.txt must_exist,must_be_versioned,install=overwrite,md5=check
@@ -216,6 +214,12 @@ File DQuasar.h must_exist,must_be_versioned,install=overwrite,md5=check
 File DServer.h must_exist,must_be_versioned,install=overwrite,md5=check
 File Base_DQuasar.h must_exist,must_be_versioned,install=overwrite,md5=check
 File Base_DServer.h must_exist,must_be_versioned,install=overwrite,md5=check
+File ASQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
+File ASServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
+File DQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
+File DServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
+File Base_DQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
+File Base_DServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
 
 Directory Meta/src install=create,
 File ASGeneralLogLevel.cpp must_exist,must_be_versioned,install=overwrite,md5=check
@@ -234,12 +238,6 @@ File Base_DSourceVariableThreadPool.cpp   		must_exist,must_be_versioned,install
 File Base_DComponentLogLevel.cpp must_exist,must_be_versioned,install=overwrite,md5=check
 File meta.cpp must_exist,must_be_versioned,install=overwrite,md5=check
 File MetaUtils.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File ASQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File ASServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File DQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File DServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File Base_DQuasar.cpp must_exist,must_be_versioned,install=overwrite,md5=check
-File Base_DServer.cpp must_exist,must_be_versioned,install=overwrite,md5=check
 
 Directory Meta/config install=create,
 File Meta.xsd must_exist,must_be_versioned,install=overwrite,md5=check
