+++ TODO.txt
@@ -1,0 +1,3 @@
+-> add nodeClasS() to UaNode() 
+--> use nodeClass() to route particular calls to addNode in NodeManager
+
